restructure file structure
