+++ src/components/pages/contact.js
@@ -3,7 +3,10 @@ import p5 from 'p5'
 import Matter from 'matter-js'
 import { connect } from 'react-redux'
 
-import { setupFrame, resetPageFrame, setupContact, 
+import { 
+  // setupFrame, 
+  // resetPageFrame, 
+  // setupContact, 
   // createMouseConstraint
  } from '../types'
 
@@ -42,7 +45,7 @@ const Contact = props => {
       const canvas = sketch.createCanvas(environment.width, environment.height)
       // createMouseConstraint(canvas.elt, engine, world, sketch)
       canvas.mouseClicked(handleClick)
-      setupFrame(environment)
+      // setupFrame(environment)
       // setupContact(environment)
     }
     sketch.draw = () => {
@@ -54,8 +57,8 @@ const Contact = props => {
       // checkGroupForMouse(environment.buttons)
     }
     sketch.windowResized = () => {
-      resetPageFrame(environment)
-      setupFrame(environment)
+      // resetPageFrame(environment)
+      // setupFrame(environment)
     }
   }
