+++ src/components/pages/home.js
@@ -3,11 +3,13 @@ import p5 from 'p5'
 import Matter from 'matter-js'
 import { connect } from 'react-redux'
 
-import { setupFrame, resetPageFrame, setupHome } from '../types'
 import {
   createMouseConstraint,
-  renderGroup
-} from '../types/utilities'
+  renderGroup,
+  setupFrame, 
+  resetPageFrame, 
+  setupHome
+} from '../utilities'
 
 const { Engine, World } = Matter
