--- src/components/types/index.js
@@ -1,4 +1,0 @@
-export * from './constructors'
-export * from './setups'
-export * from './utils'
-export * from './types'
