+++ src/components/utilities/index.ts
@@ -2,3 +2,5 @@ export * from './color'
 export * from './math'
 export * from './rendering'
 export * from './physics'
+export * from './utils'
+export * from './setups'
